pyproject.toml
[project]
name = "money"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "py-cord[speed]>=2.6.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "tortoise-orm>=0.25.1",
]
